pyproject.toml
[project]
name = "isviruspy"
version = "1.0.0"
description = "IsVirusPY - Herramienta de análisis y detección de malware"
authors = [{name = "IsVirusPY Team"}]
requires-python = ">=3.11"
dependencies = [
    "pefile>=2023.2.7",
    "lief>=0.14.0",
    "yara-python>=4.5.0",
    "scikit-learn>=1.4.0",
    "xgboost>=2.0.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "rich>=13.7.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-magic>=0.4.27",
    "requests>=2.31.0",
]

[project.scripts]
isviruspy = "isviruspy.cli:main"

[tool.setuptools]
packages = ["isviruspy", "isviruspy.modules"]
